requirements/base.txt
bioconda2biocontainer~=0.0.3
biopython~=1.77
cookiecutter~=1.7.2
cwltool~=3.0.20200710214758
docker~=4.2.2
galaxy-tool-util~=20.5.0
GitPython~=3.1.7
goenrichment~=0.0.10a3
jupyter~=1.0.0
matplotlib~=3.3.0
networkx~=2.4
numpy~=1.19.0
pandas~=1.0.5
pdf2image~=1.13.1
PyYAML~=5.3.1
requests~=2.24.0
scipy~=1.5.1
seaborn~=0.10.1
statsmodels~=0.11.1
urllib3~=1.25.9
xmltodict~=0.12.0